+++ CHANGELOG.md
@@ -22,7 +22,8 @@
 * (**SMS**) The "crop vertical borders" video setting now defaults to enabled instead of disabled; unlike the left border, the vertical borders will only ever show the current backdrop color
 * (**SNES**) In games that use the SA-1 coprocessor, the SA-1 CPU now gets a wait cycle every time it accesses SA-1 BW-RAM, similar to actual hardware
   * The SA-1 CPU still runs faster than actual hardware in some cases because bus conflict wait cycles are not emulated
-* Frontends now recognize .gen as a file extension for Genesis / Mega Drive ROM images (#149)
+* Frontends now recognize .gen and .smd as file extensions for Genesis / Mega Drive ROM images (#149)
+  * This includes attempting to auto-detect when a ROM image is interleaved (common for .smd files), and deinterleaving it during load
 * Frontends should now correctly handle files with uppercase file extensions
 
 ## Fixes